requirements.txt
defusedxml
requests
upnpy
pillow
flit
black
blacken-docs
flake8
pyproject-flake8 # for support of flake8 config pyproject.toml via pyflak8 cli
flake8-pyproject # for support of flake8 config pyproject.toml via vscode
flake8-bugbear
pylint
isort > 5.0.0
python-didl-lite
IPython
traitlets
typing_extensions
pyzmq
PySide6
qtpy
qtconsole
bleak
gpsdclient
